requirements.txt
streamlit==1.38.0
geopandas==0.14.4
python-slugify==6.1.2
requests==2.28.1
rdflib==6.1.1
owslib==0.29.3
pytest==8.3.3
